% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{clip_tiltometer}
\alias{clip_tiltometer}
\title{clip tiltometer table by date}
\usage{
clip_tiltometer(x, deploy = NA, recover = NA)
}
\arguments{
\item{x}{tibble, tiltometer}

\item{deploy}{POSIXt or NA, if not NA, clip data before this time}

\item{recover}{POSIXt or NA, if not NA, clip data before this time}
}
\value{
tibble
}
\description{
clip tiltometer table by date
}
