% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_tiltometer}
\alias{read_tiltometer}
\title{read tiltometer data file}
\usage{
read_tiltometer(filename = example_filename())
}
\arguments{
\item{filename}{character, the name of the file}
}
\value{
tibble
}
\description{
read tiltometer data file
}
