% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{tiltometer_rose}
\alias{tiltometer_rose}
\title{plot windrose for currents}
\usage{
tiltometer_rose(x)
}
\arguments{
\item{x}{tibble of current data}
}
\value{
ggplot
}
\description{
this is a wrapper around \code{\link[clifro]{windrose}}
}
